Add extraction of shortest path from start to goal
The shortest paths to all states from the initial state are already
known.  The path to the goal is pulled out of the tree simply by
rewinding from the goal to the start.

Note the returned path includes both the start and goal states, so the
number of steps to the goal is one less than the number of states in the
path.